requirements.txt
# Core Dependencies
numpy>=1.24.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0

# YOLO Object Detection
ultralytics>=8.0.0

# Computer Vision & Annotations
supervision>=0.18.0

# Tracking (BoT-SORT with Re-ID)
boxmot>=10.0.0

# Team Classification
transformers>=4.35.0
umap-learn>=0.5.0
scikit-learn>=1.3.0
Pillow>=10.0.0

# Pass Success Prediction
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
tqdm>=4.65.0

# Optional: For xT grid calculation
# (only needed if recalculating xT from raw data)
# pandas is already included above
